# Compound-Guard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CompoundGuard — compound-risk detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones/{zoneId}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get a single zone with full details
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetail"
        "404":
          description: Zone not found
  /zones/{zoneId}/sensors:
    get:
      operationId: getZoneSensors
      tags: [sensors]
      summary: Get sensor readings for a zone
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 60
          description: Number of readings to return (most recent)
        - in: query
          name: since
          schema:
            type: string
            format: date-time
          description: Return readings after this timestamp
      responses:
        "200":
          description: Sensor readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorReading"
  /zones/{zoneId}/permits:
    get:
      operationId: getZonePermits
      tags: [permits]
      summary: Get permits for a zone
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
        - in: query
          name: active_only
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Permits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permit"
  /zones/{zoneId}/alerts:
    get:
      operationId: getZoneAlerts
      tags: [alerts]
      summary: Get alerts for a zone
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /zones/{zoneId}/risk:
    get:
      operationId: getZoneRisk
      tags: [risk]
      summary: Get current risk breakdown for a zone
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskBreakdown"
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List all alerts
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: model_source
          schema:
            type: string
            enum: [single_sensor, compound]
        - in: query
          name: severity
          schema:
            type: string
            enum: [watch, elevated, critical]
        - in: query
          name: include_dismissed
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{alertId}/dismiss:
    post:
      operationId: dismissAlert
      tags: [alerts]
      summary: Dismiss an alert and update sensitivity
      parameters:
        - in: path
          name: alertId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DismissResult"
        "404":
          description: Alert not found
...
  /corpus:
    get:
      operationId: getCorpus
      tags: [corpus]
      summary: List regulatory/incident corpus entries
      parameters:
        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated tags to filter by
      responses:
        "200":
          description: Corpus entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CorpusItem"
  /corpus/{corpusId}:
    get:
      operationId: getCorpusItem
      tags: [corpus]
      summary: Get a single corpus entry
      parameters:
        - in: path
          name: corpusId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Corpus item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorpusItem"
        "404":
          description: Not found
  /shift-events:
    get:
      operationId: getShiftEvents
      tags: [shifts]
      summary: List shift events
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Shift events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftEvent"
...
  /emergency/{alertId}:
    get:
      operationId: getEmergencyResponse
      tags: [emergency]
      summary: Get auto-generated emergency draft for a critical alert
      parameters:
        - in: path
          name: alertId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emergency response draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyResponse"
        "404":
          description: Alert not found
...
components:
  schemas:
...
    Zone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hazard_class:
          type: string
          enum: [confined_space, hot_work_adjacent, electrical_hazard, gas_release_risk, general]
        polygon_points:
          type: string
          description: SVG polygon points string e.g. "100,50 200,50 200,150 100,150"
      required: [id, name, hazard_class, polygon_points]
    ZoneWithRisk:
      allOf:
        - $ref: "#/components/schemas/Zone"
        - type: object
          properties:
            current_compound_score:
              type: number
              format: float
              minimum: 0
              maximum: 100
            current_single_sensor_score:
              type: number
              format: float
              minimum: 0
              maximum: 100
            risk_level:
              type: string
              enum: [safe, watch, elevated, critical]
            active_alert_count:
              type: integer
            active_permit_count:
              type: integer
            last_sensor_at:
              type: string
              format: date-time
          required: [current_compound_score, current_single_sensor_score, risk_level, active_alert_count, active_permit_count]
    ZoneDetail:
      allOf:
        - $ref: "#/components/schemas/ZoneWithRisk"
        - type: object
          properties:
            latest_sensors:
              type: array
              items:
                $ref: "#/components/schemas/SensorReading"
            active_permits:
              type: array
              items:
                $ref: "#/components/schemas/Permit"
            active_alerts:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
            risk_breakdown:
              $ref: "#/components/schemas/RiskBreakdown"
    SensorReading:
      type: object
      properties:
        id:
          type: integer
        zone_id:
          type: integer
        timestamp:
          type: string
          format: date-time
        gas_lel_pct:
          type: number
          format: float
          description: Gas concentration as % of Lower Explosive Limit
        co_ppm:
          type: number
          format: float
          description: Carbon monoxide in parts per million
        temperature_c:
          type: number
          format: float
        vibration:
          type: number
          format: float
      required: [id, zone_id, timestamp, gas_lel_pct, co_ppm, temperature_c, vibration]
    Permit:
      type: object
      properties:
        id:
          type: integer
        zone_id:
          type: integer
        zone_name:
          type: string
        type:
          type: string
          enum: [hot_work, confined_space, electrical]
        status:
          type: string
          enum: [active, expired, cancelled]
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        issued_by:
          type: string
        has_conflict:
          type: boolean
          description: True if this permit conflicts with rising gas levels in the same zone
      required: [id, zone_id, type, status, start_time, end_time, issued_by, has_conflict]
    ShiftEvent:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        shift_name:
          type: string
          enum: [day, evening, night]
        event_type:
          type: string
          enum: [changeover, handover_note, incident_reported]
      required: [id, timestamp, shift_name, event_type]
    Alert:
      type: object
      properties:
        id:
          type: integer
        zone_id:
          type: integer
        zone_name:
          type: string
        timestamp:
          type: string
          format: date-time
        model_source:
          type: string
          enum: [single_sensor, compound]
        severity:
          type: string
          enum: [watch, elevated, critical]
        score:
          type: number
          format: float
        explanation_text:
          type: string
        citation_id:
          type: string
          nullable: true
        citation:
          $ref: "#/components/schemas/CorpusItem"
          nullable: true
        dismissed:
          type: boolean
      required: [id, zone_id, zone_name, timestamp, model_source, severity, score, explanation_text, dismissed]
    DismissResult:
      type: object
      properties:
        success:
          type: boolean
        new_sensitivity:
          type: number
          format: float
      required: [success, new_sensitivity]
...
    RiskBreakdown:
      type: object
      properties:
        zone_id:
          type: integer
        single_sensor_score:
          type: number
          format: float
        compound_score:
          type: number
          format: float
        risk_level:
          type: string
          enum: [safe, watch, elevated, critical]
        sensor_contribution:
          type: number
          format: float
          description: Fraction of compound score from sensor readings
        permit_conflict_contribution:
          type: number
          format: float
          description: Fraction of compound score from permit conflicts
        shift_context_contribution:
          type: number
          format: float
          description: Fraction of compound score from shift context
        trigger_conditions:
          type: array
          items:
            type: string
          description: Human-readable list of conditions that triggered the score
        active_citation_id:
          type: string
          nullable: true
      required: [zone_id, single_sensor_score, compound_score, risk_level, sensor_contribution, permit_conflict_contribution, shift_context_contribution, trigger_conditions]
    CorpusItem:
      type: object
      properties:
        id:
          type: string
        source_type:
          type: string
          enum: [OISD, Factory Act, Incident Summary]
        title:
          type: string
        excerpt_text:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [id, source_type, title, excerpt_text, tags]
...
    EmergencyResponse:
      type: object
      properties:
        alert_id:
          type: integer
        zone_name:
          type: string
        severity:
          type: string
        evacuation_notice:
          type: string
        incident_report:
          type: string
        generated_at:
          type: string
          format: date-time
        is_draft:
          type: boolean
      required: [alert_id, zone_name, severity, evacuation_notice, incident_report, generated_at, is_draft]
...